Updated the 24V DC/DC to output slightly above 5V
The diode-OR between the USB connector and the
24V DC/DC will allow us to choose either the USB
power or the 24V power depending on what is available.
I want the 24V DC/DC power to have higher priority
so I've increased it to 5.5V so that after the
diode-OR, it will  have a higher voltage and cause
the USB power's diode to be reverse biased.